refactor: Move, reuse `namespace`
#3068 was what started all this
